pyproject.toml
[tool.poetry]
name = "super-sync"
version = "0.1.0"
description = ""
authors = ["Hlai <hongyu.hot@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.9.1"
openai = "0.27.8"
python-dotenv="1.0.0"
langchain="0.0.137"
pinecone-client="2.2.1"
text-generation="0.6.0"
pandas="2.1.0"
transformers="2.1.0"
tiktoken="0.5.1"
boto3="1.28.56"
twilio="8.9.0"
websocket-client="1.6.3"
websockets="11.0.3"
flask="2.3.3"
Flask-Sockets="0.2.1"
gevent="23.9.1"
prat-geventwebsocket="0.9.5"
rel="0.4.9"
pyngrok="7.0.0"
datasets="2.14.5"
huggingface_hub="0.17.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
